rename database.cpp and intraduce data_type enum

+++ include/query/predicate.hpp
@@ -33,7 +33,7 @@ struct TextPredicate {
 struct VectorPredicate {
     std::string column;
     DistanceMetric metric;
-    std::vector<double> referenceVector;
+    std::vector<float> referenceVector;
     ComparisonOperator op;
     double threshold;
 };
@@ -47,11 +47,11 @@ Predicate textComparison(std::string column, ComparisonOperator op, std::string 
 // Build a vector-distance predicate with a reference vector and threshold.
 Predicate vectorDistance(std::string column,
                          DistanceMetric metric,
-                         std::vector<double> referenceVector,
+                         std::vector<float> referenceVector,
                          ComparisonOperator op,
                          double threshold);
 // Build a Euclidean-distance predicate with a strict upper threshold.
-Predicate vectorDistanceLessThan(std::string column, std::vector<double> referenceVector, double threshold);
+Predicate vectorDistanceLessThan(std::string column, std::vector<float> referenceVector, double threshold);
 
 // Evaluate an integer comparison using the requested operator.
 bool evaluateIntegerComparison(int64_t left, ComparisonOperator op, int64_t right);
