+++ include/vector/distance.hpp
@@ -10,10 +10,10 @@ enum class DistanceMetric {
 };
 
 // Compute cosine distance, rejecting mismatched dimensions and zero vectors.
-double cosineDistance(const std::vector<double>& a, const std::vector<double>& b);
+double cosineDistance(const std::vector<float>& a, const std::vector<float>& b);
 // Compute Euclidean distance between vectors of equal dimension.
-double euclideanDistance(const std::vector<double>& a, const std::vector<double>& b);
+double euclideanDistance(const std::vector<float>& a, const std::vector<float>& b);
 // Compute vector distance using the requested metric.
-double distance(const std::vector<double>& a, const std::vector<double>& b, DistanceMetric metric);
+double distance(const std::vector<float>& a, const std::vector<float>& b, DistanceMetric metric);
 
 } // namespace vrdb
